requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-filter==21.1
django-widget-tweaks==1.4.12
Pillow==9.1.0
sqlparse==0.4.2
tzdata==2022.1

